--- enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontFamilySettings/FontFamilySettings.styled.tsx
@@ -1,6 +1,0 @@
-import styled from "@emotion/styled";
-import Select from "metabase/core/components/Select";
-
-export const FontSelect = styled(Select)`
-  width: 25rem;
-`;
