--- enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontFamilySettings/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./FontFamilySettings";
