--- enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontFileSettings/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./FontFileSettings";
