--- enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontFileSettings/types.ts
@@ -1,4 +1,0 @@
-export interface FontFileOption {
-  name: string;
-  fontWeight: number;
-}
