--- enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontSettings/FontSettings.styled.tsx
@@ -1,15 +1,0 @@
-import styled from "@emotion/styled";
-import { color } from "metabase/lib/colors";
-
-export const SettingRoot = styled.div`
-  flex: 1 1 auto;
-`;
-
-export const SettingDescription = styled.div`
-  color: ${color("text-medium")};
-  margin-bottom: 1rem;
-`;
-
-export const FontFileSection = styled.div`
-  margin-top: 2rem;
-`;
