--- enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontSettings/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./FontSettings";
