+++ enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontSettingsWidget/utils.ts
